Moved TagInfoWrapper file and updated imports
