+++ src/pages/TagInfo/TagInfoWrapper.tsx
@@ -1,9 +1,8 @@
 import { TagInfo, AsyncStatus } from '@equinor/procosys-webapp-components';
 import axios from 'axios';
-import React, { useState } from 'react';
-import { useEffect } from 'react';
-import { AdditionalTagField, TagDetails } from '../services/apiTypes';
-import useCommonHooks from '../utils/useCommonHooks';
+import React, { useState, useEffect } from 'react';
+import { AdditionalTagField, TagDetails } from '../../services/apiTypes';
+import useCommonHooks from '../../utils/useCommonHooks';
 
 type TagInfoWrapperProps = {
     tagId?: number;
@@ -40,13 +39,11 @@ const TagInfoWrapper = ({ tagId }: TagInfoWrapperProps): JSX.Element => {
     }, [tagId]);
 
     return (
-        <>
-            <TagInfo
-                tagInfo={tagInfo}
-                fetchTagStatus={fetchTagStatus}
-                additionalFields={additionalFields}
-            />
-        </>
+        <TagInfo
+            tagInfo={tagInfo}
+            fetchTagStatus={fetchTagStatus}
+            additionalFields={additionalFields}
+        />
     );
 };
